Add files via upload
Online retail is a transnational data set that contains all the transactions occurring between 01/12/2010 and 09/12/2011 for a UK-based and registered non-store online retail. The company mainly sells unique all-occasion gifts. Many customers of the company are wholesalers.
We will be using the online retail transactional dataset to build an RFM (R(Recency), F(Frequency), M(Monetary)) clustering and choose the best set of customers.